move the triangle color logic into helper
